Kafka Improvements

+++ investment-service/src/main/java/com/rahim/investmentservice/service/holding/HoldingUpdateService.java
@@ -1,10 +1,12 @@
 package com.rahim.investmentservice.service.holding;
 
+import java.math.BigDecimal;
+
 /**
  * @author Rahim Ahmed
  * @created 27/05/2024
  */
 public interface HoldingUpdateService {
 
-    void updateCurrentValue(Integer goldTypeId, Double updatedPrice);
+    void updateCurrentValue(Integer goldTypeId, BigDecimal updatedPrice);
 }
